+++ CHANGELOG.md
@@ -20,6 +20,9 @@
 - `rename schemas.list_schemas to schemas.schemas_list`
 - `rename operations.list_queries to operations.queries_list`
 - `rename operations.mutations_list to operation.mutations_list`
+- `rename types.list_types to types.types_list`
+- `map lua 'integer' to GraphQL 'Long' type`
+- `updated description of all APIs`
 
 ## 0.0.1